fix(item): 修正 code review 找出的三個正確性問題
fresh-context code review 對 7737dcd 的 diff 找出 4 筆 findings，本筆處理其中
三筆正確性問題（第 4 筆為 N+1 效能議題，PLAUSIBLE 且目前衝擊接近零，不在本筆範圍）。

1. ItemService.amend 的變更偵測繞過審核護欄（medium）
   contentChanged 原本以正規化後的名稱判斷有無變更，但 displayName 是無條件寫回的
   獨立欄位。正規化會剝掉所有標點、符號、空白與全半形差異，因此對一個 VERIFIED 節點
   送 displayName="W-i-F-i 模.組!!! ★★★"，normalizedName 不變、判定為「無變更」而不
   退回 DRAFT，displayName 卻照樣寫進去——已驗證資料的顯示名稱可被任意改寫且不留審核
   痕跡，與 data-provenance spec「已驗證的資料 MUST NOT 被靜默改寫」直接衝突。
   改為比對 displayName 原字串；正規化仍只用於名稱衝突檢查。

2. 分頁位移量整數溢位（low）
   page 只有 @Min(0) 沒有上限，page * size 以 int 相乘在大頁碼時溢位成負數，送出
   負的 OFFSET 讓 PostgreSQL 報錯，端點回 500 而非空頁。改以 long 計算，
   ItemRepository.findEndProducts 的 offset 參數型別一併放大。

3. 純標點的搜尋關鍵字回 400（low）
   關鍵字沿用必填名稱的 NameNormalizer.normalize，name=--- 這類輸入正規化後為空而
   拋 400「名稱不可為空」，與規格要求的「無符合資料回空清單」不符，也與 name 全為
   空白時回全部的既有行為不一致。NameNormalizer 拆出不拋例外的 normalizeOrEmpty，
   normalize 改為委派給它，正規化規則仍只有一份；空結果視為不過濾。

測試：
- ItemServiceTest 補 3 筆（標點差異仍退回 DRAFT 且寫回、純標點關鍵字不過濾、
  大頁碼位移量不溢位）；amend_sameValues_shouldKeepVerified 的語意隨規格收緊，
  由「正規化後相同」改為「逐字元相同」。
- 全量 ./mvnw clean verify 通過：surefire 163、failsafe 68、0 失敗。

文件：
- design.md D3 補修訂註記：原文寫「正規化後比較 displayName」是這個洞的來源，
  改以 spec 為準並記下另外兩個被否決方向的理由。
- CHANGELOG 的無變更判準改寫為「三個欄位逐字元相同」。

+++ src/main/java/com/profetai/industrymap/repository/ItemRepository.java
@@ -38,7 +38,7 @@ public interface ItemRepository extends JpaRepository<Item, Long> {
     List<Item> findEndProducts(@Param("reviewStatuses") Collection<String> reviewStatuses,
                                @Param("namePattern") String namePattern,
                                @Param("limit") int limit,
-                               @Param("offset") int offset);
+                               @Param("offset") long offset);
 
     /** 終端成品列表的總筆數，過濾條件與 {@link #findEndProducts} 一致 */
     @Query(value = """
